refactor(cli): extract symlink classification into src/cli/
main.rs grew from 1148 to 1508 lines, well past the 500-600 guideline in
AGENTS.md, and the symlink logic measured ~110 code lines against the
plan's 60-line extraction threshold.

src/cli/symlink.rs takes the symlink-specific classification;
src/cli/mod.rs takes FileOutcome, stdout_is_terminal and synthetic_result,
which are not symlink-specific -- the directory path uses synthetic_result
too, and FileOutcome is process_file's contract. Unit tests move with the
code they cover.

Pure code movement: no behavior change, and the U1-U5 suites pass
unchanged. main.rs is back to 1271 lines.

Implements U6 of the issue #383 plan.
